+++ src/Sockets.hpp
@@ -14,14 +14,14 @@ class Sockets
 {
 	public:
 
-						Sockets(int socket_limit);
+						Sockets(size_t socket_limit);
 						~Sockets(void);
 		void			close(socket_t &socket);
 		void			close_all(void);
 		void			close_epoll(void);
 		int				add(socket_t &socket);
-		int				limit(void) const;
-		int				size(void) const;
+		size_t			limit(void) const;
+		size_t			size(void) const;
 		socket_t		&at(int i);
 		int				epoll_inst(void) const;
 		epoll_event		*events_addr(void);
@@ -31,8 +31,8 @@ class Sockets
 	private:
 
 		EpollManager			m_epoll;
-		const int				m_socket_limit;
-		int						m_size;
+		const size_t			m_socket_limit;
+		size_t					m_size;
 		std::vector<socket_t>	m_sockets;
 
 	// DEBUG
